+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - Some files were changed and fixed because some lines were duplicated.
 - Added the option to add hotwired and stimulus when creating a new project. This is optional.
+- Some rules of prettier and eslint were added to the eslintrc.json file.